Extract Router::RegistryLookup (verbatim move, no behavior change)
NxN G14 3c, R12: code-move only. Move the eight private registry lookup
helpers (registry_tier, registry_tier_for_default_provider,
registry_provider_for_tier, registry_entry_for_provider,
registry_model_for_tier, registry_entry_for_tier, registry_default_model,
registry_resolution_metadata) out of Router's 1030-line body into a
Router::RegistryLookup mixin, extended into the Router singleton.

The module is lexically nested under Router so PROVIDER_TIER /
PROVIDER_ORDER / Call::Registry / Resolution / log / handle_exception
resolve identically; extend keeps every method on the same self with the
same private visibility, so cross-calls and ivars are unchanged. No
external caller references these private methods (grep-verified across
lib + spec). Full rspec 3219/0, rubocop 489/0.
